+++ src/Quests/CustomQuest.cs
@@ -16,7 +16,7 @@ namespace QuestFramework.Quests
     /// <summary>
     /// Custom quest definition
     /// </summary>
-    public class CustomQuest
+    public class CustomQuest : IDisposable
     {
         private int _customTypeId;
         private string _trigger;
@@ -49,6 +49,11 @@ namespace QuestFramework.Quests
         public QuestLogColors Colors { get; set; }
         public List<Hook> Hooks { get; set; }
         public Dictionary<string, int> FriendshipGain { get; }
+
+        public virtual void OnRegister()
+        {
+        }
+
         public Dictionary<string, string> Tags { get; }
 
         public string Name
@@ -104,6 +109,11 @@ namespace QuestFramework.Quests
             this.Tags = new Dictionary<string, string>();
             this._defaultObjective = new CustomQuestObjective("default", "");
             this._currentObjectives = new List<CustomQuestObjective>();
+            this.OnInitialize();
+        }
+
+        protected virtual void OnInitialize()
+        {
         }
 
         public CustomQuest(string name) : this()
@@ -133,6 +143,11 @@ namespace QuestFramework.Quests
             }
         }
 
+        public void ForceUpdate()
+        {
+            this.NeedsUpdate = true;
+        }
+
         internal void ConfirmAccept(IQuestInfo questInfo)
         {
             StatsManager.AddAcceptedQuest(this.GetFullName());
@@ -150,6 +165,14 @@ namespace QuestFramework.Quests
         /// (NeedsUpdate field is set to TRUE)
         /// </summary>
         internal virtual void Update()
+        {
+            this.OnUpdate();
+        }
+
+        /// <summary>
+        /// Do some things when quest is updating
+        /// </summary>
+        protected virtual void OnUpdate()
         {
         }
 
@@ -233,6 +256,17 @@ namespace QuestFramework.Quests
                 return name;
 
             return $"{name}@{this.OwnedByModUid}";
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+        }
+
+
+        public void Dispose()
+        {
+            this.Dispose(disposing: true);
+            GC.SuppressFinalize(this);
         }
     }
 }