+++ src/Quests/CustomQuest.stateful.cs
@@ -21,6 +21,7 @@ namespace QuestFramework.Quests
         /// A state data
         /// </summary>
         public TState State { get; private set; }
+        public bool NeedsSync { get; private set; }
 
         /// <summary>
         /// Create custom quest with state (Statefull quest)
@@ -46,6 +47,12 @@ namespace QuestFramework.Quests
         /// CALL THIS METHOD EVER WHEN YOU CHANGED QUEST STATE!
         /// </summary>
         public void Sync()
+        {
+            this.NeedsSync = true;
+            this.NeedsUpdate = true;
+        }
+
+        public void PerformSync()
         {
             var payload = new StatePayload(
                 questName: this.GetFullName(),
@@ -113,9 +120,9 @@ namespace QuestFramework.Quests
         {
             base.Update();
 
-            if (this.State is IReactiveState activeState && activeState.WasChanged)
+            if (this.NeedsSync || this.State is IReactiveState activeState && activeState.WasChanged)
             {
-                this.Sync();
+                this.PerformSync();
             }
         }
 
@@ -152,7 +159,7 @@ namespace QuestFramework.Quests
             if (this.State is IPersistentState resetableState)
             {
                 resetableState.Reset();
-                this.Sync();
+                this.PerformSync();
                 return;
             }
 
@@ -162,7 +169,7 @@ namespace QuestFramework.Quests
             }
 
             this.State = this.PrepareState();
-            this.Sync();
+            this.PerformSync();
         }
 
         /// <summary>